[Sanitizer] Add smoke test for report symbolication
Add a test that checks symbolication for out-of-process and in-process
symbolication with and without debug info.

Note that there already exist tests in the compiler-rt test suite that
check report symbolication.  The purpose of this test is to have a
simple smoke test for Swift, and specifically Swift demangling.

rdar://62753845